adjust task return for dumps

+++ pipelines/rj_smtr/br_rj_riodejaneiro_rdo/tasks.py
@@ -71,7 +71,7 @@ from pipelines.utils.utils import log
 
 @task
 def get_file_paths_from_ftp(
-    transport_mode: str, report_type: str, wait=None
+    transport_mode: str, report_type: str, wait=None, dump=False
 ):  # pylint: disable=W0613
     """
     Search for files inside previous interval (days) from current date,
@@ -80,12 +80,17 @@ def get_file_paths_from_ftp(
 
     execution_time = pendulum.now(constants.TIMEZONE.value)
     # Define interval for date search
-    now = execution_time + timedelta(hours=11, minutes=30)
-    min_timestamp = (
-        now - timedelta(days=1)
-    ).timestamp()  # Saturday, January 8, 2022 2:30:00 PM
-    max_timestamp = now.timestamp()  # Sunday, January 9, 2022 2:30:00 PM
-    log(f"{execution_time} of type {type(execution_time)}")
+    now = execution_time + timedelta(
+        hours=11, minutes=30
+    )  # Sunday, January 9, 2022 2:30:00 PM
+    max_timestamp = now.timestamp()
+    if not dump:
+        min_timestamp = (
+            now - timedelta(days=1)
+        ).timestamp()  # Saturday, January 8, 2022 2:30:00 PM
+        log(f"{execution_time} of type {type(execution_time)}")
+    else:
+        min_timestamp = datetime(2022, 1, 1).timestamp()
     # Connect to FTP & search files
     ftp_client = connect_ftp()
     files_updated_times = {
@@ -93,9 +98,10 @@ def get_file_paths_from_ftp(
         for file, info in ftp_client.mlsd(transport_mode)
     }
     # Get files modified inside interval
+    files = []
     for filename, file_mtime in files_updated_times.items():
         if min_timestamp <= file_mtime < max_timestamp:
-            if filename[:3] == report_type:
+            if filename[:3] == report_type and "HISTORICO" not in filename:
                 log(
                     f"""
                     Found file
@@ -116,7 +122,8 @@ def get_file_paths_from_ftp(
                     "partitions": f"ano={date[:4]}/mes={date[4:6]}/dia={date[6:]}",
                 }
                 log(f"Create file info: {file_info}")
-    return file_info
+                files.append(file_info)
+    return files
 
 
 @task
@@ -145,11 +152,12 @@ def download_and_save_local_from_ftp(file_info: dict):
     Path(file_info["raw_path"]).parent.mkdir(parents=True, exist_ok=True)
     # Get data from FTP - TODO: create get_raw() error alike
     ftp_client = connect_ftp()
-    with open(file_info["raw_path"], "wb") as raw_file:
-        ftp_client.retrbinary(
-            "RETR " + file_info["ftp_path"],
-            raw_file.write,
-        )
+    if not Path(file_info["raw_path"]).is_file():
+        with open(file_info["raw_path"], "wb") as raw_file:
+            ftp_client.retrbinary(
+                "RETR " + file_info["ftp_path"],
+                raw_file.write,
+            )
     ftp_client.quit()
     # Get timestamp of download time
     file_info["timestamp_captura"] = pendulum.now(constants.TIMEZONE.value).isoformat()
@@ -159,11 +167,11 @@ def download_and_save_local_from_ftp(file_info: dict):
     return file_info
 
 
-@task
+@task(nout=3)
 def pre_treatment_br_rj_riodejaneiro_rdo(
-    file_info: dict,
+    files: list,
     divide_columns_by: int = 100,
-):
+) -> tuple:
     """Adds header, capture_time and standardize columns
 
     Args:
@@ -174,49 +182,60 @@ def pre_treatment_br_rj_riodejaneiro_rdo(
     Returns:
         dict: updated file_info with treated filepath
     """
-    config = rdo_constants.RDO_PRE_TREATMENT_CONFIG.value[file_info["transport_mode"]][
-        file_info["report_type"]
-    ]
-    # context.log.info(f"Config for ETL: {config}")
-    # Load data
-    df = pd.read_csv(  # pylint: disable=C0103
-        file_info["raw_path"], header=None, delimiter=";", index_col=False
-    )  # pylint: disable=C0103
-    log(f"Load csv from raw file:\n{df.head(5)}")
-    # Set column names for those already in the file
-    df.columns = config["reindex_columns"][: len(df.columns)]
-    log(f"Found reindex columns at config:\n{df.head(5)}")
-    # Treat column "codigo", add empty column if doesn't exist
-    if ("codigo" in df.columns) and (file_info["transport_mode"] == "STPL"):
-        df["codigo"] = df["codigo"].str.extract("(?:VAN)(\\d+)").astype(str)
-    else:
-        df["codigo"] = ""
-    # Order columns
-    # if config["reorder_columns"]:
-    #     ordered = [
-    #         config["reorder_columns"][col]
-    #         if col in config["reorder_columns"].keys()
-    #         else i
-    #         for i, col in enumerate(config["reindex_columns"])
-    #     ]
-    #     df = df[list(config["reindex_columns"][col] for col in ordered)] # pylint: disable=C0103
-    # else:
-    df = df[config["reindex_columns"]]
-    # Add timestamp column
-    df["timestamp_captura"] = file_info["timestamp_captura"]
-    log(f"Added timestamp_captura: {file_info['timestamp_captura']}")
-    log(f"Before dividing df is\n{df.head(5)}")
-    # Divide columns by value
-    if config["divide_columns"]:
-        df[config["divide_columns"]] = df[config["divide_columns"]].apply(
-            lambda x: x / divide_columns_by, axis=1
-        )
-    # Save treated data
-    file_info["treated_path"] = file_info["local_path"].format(
-        bucket_mode="staging", file_ext="csv"
-    )
-    Path(file_info["treated_path"]).parent.mkdir(parents=True, exist_ok=True)
-    df.to_csv(file_info["treated_path"], index=False)
-    log(f'Saved treated data to: {file_info["treated_path"]}')
-    log(f"Updated file info is:\n{file_info}")
-    return file_info
+    treated_paths, raw_paths, partitions = [], [], []
+    for file_info in files:
+        try:
+            config = rdo_constants.RDO_PRE_TREATMENT_CONFIG.value[
+                file_info["transport_mode"]
+            ][file_info["report_type"]]
+            # context.log.info(f"Config for ETL: {config}")
+            # Load data
+            df = pd.read_csv(  # pylint: disable=C0103
+                file_info["raw_path"], header=None, delimiter=";", index_col=False
+            )  # pylint: disable=C0103
+            log(f"Load csv from raw file:\n{df.head(5)}")
+            # Set column names for those already in the file
+            df.columns = config["reindex_columns"][: len(df.columns)]
+            log(f"Found reindex columns at config:\n{df.head(5)}")
+            # Treat column "codigo", add empty column if doesn't exist
+            if ("codigo" in df.columns) and (file_info["transport_mode"] == "STPL"):
+                df["codigo"] = df["codigo"].str.extract("(?:VAN)(\\d+)").astype(str)
+            else:
+                df["codigo"] = ""
+            # Order columns
+            # if config["reorder_columns"]:
+            #     ordered = [
+            #         config["reorder_columns"][col]
+            #         if col in config["reorder_columns"].keys()
+            #         else i
+            #         for i, col in enumerate(config["reindex_columns"])
+            #     ]
+            #     df = df[list(config["reindex_columns"][col] for col in ordered)]
+            # # pylint: disable=C0103
+            # else:
+            df = df[config["reindex_columns"]]
+            # Add timestamp column
+            df["timestamp_captura"] = file_info["timestamp_captura"]
+            log(f"Added timestamp_captura: {file_info['timestamp_captura']}")
+            log(f"Before dividing df is\n{df.head(5)}")
+            # Divide columns by value
+            if config["divide_columns"]:
+                df[config["divide_columns"]] = df[config["divide_columns"]].apply(
+                    lambda x: x / divide_columns_by, axis=1
+                )
+            # Save treated data
+            file_info["treated_path"] = file_info["local_path"].format(
+                bucket_mode="staging", file_ext="csv"
+            )
+            Path(file_info["treated_path"]).parent.mkdir(parents=True, exist_ok=True)
+            df.to_csv(file_info["treated_path"], index=False)
+            log(f'Saved treated data to: {file_info["treated_path"]}')
+            log(f"Updated file info is:\n{file_info}")
+            # Build returns
+            treated_paths.append(file_info["treated_path"])
+            raw_paths.append(file_info["raw_path"])
+            partitions.append(file_info["partitions"])
+        except Exception as e:
+            log(f"Pre Treatment failed with error: {e}")
+            continue
+    return treated_paths, raw_paths, partitions
